# [AArch64] Optimized memcmp