@lineageos-gerrit 2017-12-21
Hi there;
This is an automated reply, to let you know LineageOS doesn't merge pull
requests submitted directly to our github.com repositories.

You can submit your change requests through our public review system,
located at https://review.lineageos.org/

That review system is an installation of Google's Gerrit code review.
If you've never interacted with Gerrit before, we also suggest you take
a look at http://wiki.lineageos.org/usinggerrit-howto.html

Thanks!